requirements.txt
black==23.9.1
pylint==3.0.0
pytest==7.4.2
pytest-cov==4.1.0
python-multipart==0.0.6
uvicorn==0.23.2
fastapi==0.103.2
Pillow==10.0.1
torchvision==0.15.2

